Reorganize project structure into packages
