+++ docs/README.md
@@ -1,4 +1,4 @@
-# B33PBOP User Guide
+# B33PBOP.ui.B33PBOP User Guide
 
 // Update the title above to match the actual product name
