+++ src/main/java/B33PBOP/command/ByeCommand.java
@@ -1,3 +1,7 @@
+package B33PBOP.command;
+
+import B33PBOP.ui.UI;
+
 public class ByeCommand implements CommandExecutor {
     private final UI UI;
