+++ src/main/java/B33PBOP/command/CommandExecutor.java
@@ -1,3 +1,7 @@
+package B33PBOP.command;
+
+import B33PBOP.exception.BotException;
+
 public interface CommandExecutor {
     boolean execute(String arg) throws BotException;
 }
