+++ src/main/java/B33PBOP/command/DeleteTaskCommand.java
@@ -1,3 +1,10 @@
+package B33PBOP.command;
+
+import B33PBOP.exception.BotException;
+import B33PBOP.storage.Storage;
+import B33PBOP.task.Task;
+import B33PBOP.ui.UI;
+import B33PBOP.task.TaskList;
 import java.io.IOException;
 
 public class DeleteTaskCommand implements CommandExecutor {
