+++ src/main/java/B33PBOP/exception/BotException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class BotException extends Exception {
+    public BotException(String msg) {
+        super(msg);
+    }
+}
