+++ src/main/java/B33PBOP/exception/IncompleteArgumentException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class IncompleteArgumentException extends BotException {
+    public IncompleteArgumentException(String msg) {
+        super(msg);
+    }
+}
