+++ src/main/java/B33PBOP/exception/InvalidArgumentException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class InvalidArgumentException extends BotException {
+    public InvalidArgumentException(String msg) {
+        super(msg);
+    }
+}
