+++ src/main/java/B33PBOP/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class InvalidCommandException extends BotException {
+    public InvalidCommandException(String msg) {
+        super(msg);
+    }
+}
