+++ src/main/java/B33PBOP/exception/InvalidEventEndDateException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class InvalidEventEndDateException extends BotException {
+    public InvalidEventEndDateException(String msg) {
+        super(msg);
+    }
+}
