+++ src/main/java/B33PBOP/exception/TaskListIndexOutOfBoundException.java
@@ -1,0 +1,7 @@
+package B33PBOP.exception;
+
+public class TaskListIndexOutOfBoundException extends BotException {
+    public TaskListIndexOutOfBoundException(String msg) {
+        super(msg);
+    }
+}
