+++ src/main/java/B33PBOP/storage/Storage.java
@@ -1,15 +1,18 @@
+package B33PBOP.storage;
+
+import B33PBOP.task.Task;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.List;
 
 /**
- * Storage handles storage for B33PBOP.
- * On first usage, it creates a new DIRECTORY with a file B33PBOP.txt which stores tasks in String format.
+ * B33PBOP.storage.Storage handles storage for B33PBOP.ui.B33PBOP.
+ * On first usage, it creates a new DIRECTORY with a file B33PBOP.ui.B33PBOP.txt which stores tasks in String format.
  * Subsequent usage overwrites the existing file to update storage.
  */
 public class Storage {
-    private final File DIRECTORY;
     private final File STORAGE_FILE;
 
     /**
@@ -18,10 +21,10 @@ public class Storage {
      * @throws IOException If storage creation fails
      */
     public Storage() throws IOException {
-        DIRECTORY = new File("data");
+        File DIRECTORY = new File("data");
         DIRECTORY.mkdirs();
 
-        STORAGE_FILE = new File(DIRECTORY, "B33PBOP.txt");
+        STORAGE_FILE = new File(DIRECTORY, "B33PBOP.ui.B33PBOP.txt");
         if (!STORAGE_FILE.exists()) {
             boolean isCreated = STORAGE_FILE.createNewFile();
             if (!isCreated) {
@@ -35,7 +38,7 @@ public class Storage {
     }
 
     /**
-     * Updates B33PBOP.txt file with a new task list.
+     * Updates B33PBOP.ui.B33PBOP.txt file with a new task list.
      * @param tasks List of tasks added by the user.
      * @throws IOException If update fails.
      */
