+++ src/main/java/B33PBOP/util/DateTimeParser.java
@@ -1,3 +1,5 @@
+package B33PBOP.util;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
