--- src/main/java/BotException.java
@@ -1,35 +1,0 @@
-public class BotException extends Exception {
-    public BotException(String msg) {
-        super(msg);
-    }
-}
-
-class InvalidCommandException extends BotException {
-    public InvalidCommandException(String msg) {
-        super(msg);
-    }
-}
-
-class InvalidArgumentException extends BotException {
-    public InvalidArgumentException(String msg) {
-        super(msg);
-    }
-}
-
-class IncompleteArgumentException extends BotException {
-    public IncompleteArgumentException(String msg) {
-        super(msg);
-    }
-}
-
-class TaskListIndexOutOfBoundException extends BotException {
-    public TaskListIndexOutOfBoundException(String msg) {
-        super(msg);
-    }
-}
-
-class InvalidEventEndDateException extends BotException {
-    public InvalidEventEndDateException(String msg) {
-        super(msg);
-    }
-}
