+++ text-ui-test/tests/deadline_expected.txt
@@ -1,5 +1,5 @@
 ___________________________________________________________________________
-I'm B33PBOP...
+I'm B33PBOP.ui.B33PBOP...
 What do you want?
 ___________________________________________________________________________
